Move the fix inside `Batch.readMessage`
The reasoning is that `Batch.readMessage` is normally where we update
the batch offset so the goal is to do this additional update to the
batch offset in the same location.

Also adjusts the test case retry so that the test case may finish
earlier than the previous sleep time while also giving the test more
time to finish to reduce flakiness.

The test case also now ensures that the final state is fully compacted
before declaring success.